refactor(chess/application/web): Controller에서 Application Annotation 분리
Application 계층 분리하며 실행 계층 class rename
- Application -> Controller
- Controller -> Service
