+++ src/main/java/chess/application/web/spring/GameController.java
@@ -1,0 +1,84 @@
+package chess.application.web.spring;
+
+import chess.application.web.JsonTransformer;
+import chess.application.web.GameService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.SpringApplication;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+public class GameController {
+    @Autowired
+    private GameService gameService;
+    private final JsonTransformer jsonTransformer;
+
+    public GameController() {
+        this.jsonTransformer = new JsonTransformer();
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(GameController.class, args);
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        model.addAllAttributes(gameService.modelReady());
+        return "index";
+    }
+
+    @GetMapping("/start")
+    public String start(Model model) {
+        gameService.start();
+        return play(model);
+    }
+
+    @GetMapping("/load")
+    public String load(Model model) {
+        gameService.load();
+        return play(model);
+    }
+
+    @PostMapping("/move")
+    public String move(Model model, @RequestParam String source, @RequestParam String target) {
+        gameService.move(source, target);
+        if (gameService.isGameFinished()) {
+            return end(model);
+        }
+        return play(model);
+    }
+
+    private String play(Model model) {
+        model.addAllAttributes(gameService.modelPlayingBoard());
+        return "index";
+    }
+
+    @GetMapping("/status")
+    public ResponseEntity<String> status() {
+        String statusData = jsonTransformer.render(gameService.modelStatus());
+        return ResponseEntity.ok().body(statusData);
+    }
+
+    @GetMapping("/save")
+    public ResponseEntity<Void> save() {
+        gameService.save();
+        return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
+
+    @GetMapping("/end")
+    public String end(Model model) {
+        model.addAllAttributes(gameService.end());
+        return "result";
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<String> handle(Exception exception) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("[ERROR] " + exception.getMessage());
+    }
+}
